pyproject.toml
[project]
name = "car-gui"
version = "0.1.0"
description = "Car Management System with PyQT6 & sqlite"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyqt6-tools>=6.4.2.3.3",
]
